+++ R/show_downsample_curves.R
@@ -9,38 +9,36 @@
 #'
 #' @examples show_downsample_curves(3000, 5000, 10000)
 #'
-#' @value Called for side effects.
-#'
-#' TODO -- update parameter names in downsample, downsample_spectra
+#' @return Called for side effects.
 
 show_downsample_curves <- function(...) {
-  thresholds <- unlist(list(...))
-  last <- length(thresholds)
+  th <- unlist(list(...))
+  last <- length(th)
   stopifnot(last >= 1)
-  maxt <- max(thresholds)
+  maxt <- max(th)
   maxx <- maxt * 1.5
-  mint <- min(thresholds)
+  mint <- min(th)
   xvec <- round(seq(0, maxx, by = maxx / 1000))
   pch <- "."
   lty <- 3
-  start.line.fac <- 0.5
-  thresholds <- sort(thresholds)
-  plot(xvec, downsample(xvec, thres = maxt),
+  line.start.x <- maxx * 0.08
+  th <- sort(th)
+  plot(xvec, downsample(xvec, downsample_threshold = maxt),
        pch = pch,
        xlab = "Original number of mutations",
-       ylab = "Number of mutations after downsampling")
-  lines(x = c(mint * start.line.fac, maxx), y = rep(maxt, 2), type="l", lty = lty)
+       ylab = "Number of mutations after downsampling",
+       ylim = c(0, 1.1 * maxt))
+  lines(x = c(line.start.x, maxx),
+        y = rep(maxt, 2), type="l", lty = lty)
   text(x = 0, y = maxt, maxt, adj = c(0, 0.5))
   text(x = 0, y = maxt * 1.07, "downsampling_threshold", adj = c(0, 0.5))
-  # abline(h = thresholds[last], lty = lty)
-  if (length(thresholds) > 1) {
-    thresholds <- thresholds[-last]
-    for (tt in thresholds) {
-      # browser()
-      points(xvec, downsample(xvec, thres = tt), pch = pch, new = FALSE)
-      lines(x = c(mint * start.line.fac, maxx), y = rep(tt, 2), type="l", lty = lty)
+  if (length(th) > 1) {
+    th <- th[-last]
+    for (tt in th) {
+      points(xvec, downsample(xvec, downsample_threshold = tt), pch = pch, new = FALSE)
+      lines(x = c(line.start.x, maxx),
+            y = rep(tt, 2), type="l", lty = lty)
       text(x = 0, y = tt, tt, adj = c(0, 0.5))
-      # abline(h = tt, lty = lty)
     }
   }
 }